whiteboard/requirements.txt
alembic==1.0.0
aniso8601==3.0.2
asn1crypto==0.24.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.3.1
dnspython==1.15.0
eventlet==0.24.1
Flask==1.0.2
Flask-Admin==1.5.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.6
Flask-Login==0.4.1
Flask-Migrate==2.2.1
Flask-RESTful==0.3.6
Flask-SocketIO==3.0.2
Flask-SQLAlchemy==2.3.2
greenlet==0.4.15
gunicorn==19.9.0
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
monotonic==1.5
pycparser==2.19
PyJWT==1.6.4
PyMySQL==0.9.2
python-dateutil==2.7.3
python-editor==1.0.3
python-engineio==2.3.2
python-socketio==2.0.0
pytz==2018.5
requests==2.21.0
six==1.11.0
SQLAlchemy==1.2.12
stripe==2.18.1
urllib3==1.23
uWSGI==2.0.17.1
Werkzeug==0.14.1
WTForms==2.2.1
